backend/requirements.txt
asgiref==3.5.2
distlib==0.3.4
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
PyJWT
djongo==1.3.6
dnspython==2.2.1
filelock==3.7.0
platformdirs==2.5.2
pymongo==3.12.3
pytz==2022.1
six==1.16.0
sqlparse==0.2.4
tzdata==2022.1
gunicorn